Do minor refactoring
